requirements/local.txt
-r base.txt
beautifulsoup4
fabric
gitpython
hgapi
ipdb
psycopg2
pytest-django
pyyaml
south
walkdir
